Move conflicts to the term comparison
Just by moving these we fix the 'existing statement, started during a
term' specs and allows the 'Exact match when there is no actions needed'
change to happen without specs failing.
